Lambdaのデプロイコマンドの整理を行う

+++ utils.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 
-# 名前のデータ取る
 name=$(cat yourname.env | sed s/YOURNAME=//)
 
 writename(){
@@ -17,34 +16,56 @@ writename(){
     fi
 }
 
-lambda_deploy(){
-
+lambda_deploy() {
+        
+    # ファイル名が未指定の場合は処理を中断します。
+    if [ $# -lt 1 ]; then
+        echo "デプロイしたいファイル名を指定しましょう Usage: ./utils.sh lambda_deploy <FILENAME>"
+        exit 1
+    fi
+    
+    filename=$1
+    if [ "$(ls lambda/ | grep -x "${filename}")" ]; then
+        echo "関数${filename}の処理を開始します"
+    else
+        # 指定したファイル名が存在しない場合は処理を中断します。
+        echo "${filename}はlambda/配下に存在しないようです 存在するファイル名を指定しましょう"
+        exit 1
+    fi
+    
     cd lambda
-
-    for lambdaf in *.js
-    do
-        funame=$(echo $lambdaf | sed s/.js//)
-        echo 関数${funame}のデプロイを開始します
-        zip -r ${funame}.zip ${lambdaf}
+    funame=$(echo $filename | sed s/.js//)
+    aws lambda get-function --function-name $funame > /dev/null 2>&1
+    if [ $? -eq 0 ]; then
+        # functionが存在している時
+        echo "関数${funame}の内容を更新します"
+        zip -r ${funame}.zip ${filename}
+        aws lambda update-function-code  --function-name ${funame} \
+        --zip-file fileb://${funame}.zip
+        rm ${funame}.zip
+            
+    else
+        # functionが存在していない時
+        echo "関数${funame}のデプロイを開始します"
+        zip -r ${funame}.zip ${filename}
         aws lambda create-function  --function-name ${funame} \
         --runtime nodejs16.x \
         --role arn:aws:iam::347867041416:role/internship_exec_role \
         --handler ${funame}.handler  --zip-file fileb://${funame}.zip \
         --region ap-northeast-1
         rm ${funame}.zip
-    done
+    fi
     cd ..
+    
 }
 
-lambda_update(){
-    cd lambda
-    for updatelist in *.js
+lambda_deploy_all(){
+
+    for lambdaf in lambda/*.js
     do
-        funame=$(echo $updatelist | sed s/.js//)
-        echo ${funame}の更新
-        zip -r ${funame}.zip ${updatelist}
-        aws lambda update-function-code --function-name ${funame} --zip-file fileb://${funame}.zip
-        rm ${funame}.zip
+        # lambda_deployはファイル名を引数として受け取る想定のため、不要な"lambda/"を取り除いています
+        filename=$(echo $lambdaf | sed "s/lambda\///")
+        lambda_deploy $filename
     done
     cd ..
 }
@@ -99,8 +120,8 @@ deletename(){
 
 case $1 in
     "") writename;;
-    lambda_deploy) lambda_deploy;;
-    lambda_update) lambda_update;;
+    lambda_deploy_all) lambda_deploy_all;;
+    lambda_deploy) lambda_deploy $2;;
     s3_deploy) s3_deploy;;
     s3_update) s3_update;;
     delete_name) deletename;;
